+++ cypress/locators/index.js
@@ -10,22 +10,28 @@ import {
   LABEL,
   DLS_ROOT,
   WARNING_ICON,
+  INFO_ICON,
   COMMON_INPUT_PREFIX,
   COMMON_INPUT_CHARACTER_LIMIT,
   CY_ROOT,
   PORTAL,
+  LEGEND,
   STICKY_FOOTER,
 } from "./locators";
 
 // component preview locators
 export const commonButtonPreviewRoot = () => cy.get(DLS_ROOT).find("button");
 export const label = () => cy.get(LABEL);
+export const legend = () => cy.get(LEGEND);
+export const legendSpan = () => cy.get(`legend > span`);
 export const backgroundUILocator = () => cy.get(BACKGROUND_UI_LOCATOR);
 
 // component preview locators
 export const dlsRoot = () => cy.get(DLS_ROOT);
 export const icon = () => cy.get(ICON);
+export const errorIcon = () => cy.get(ICON);
 export const warningIcon = () => cy.get(WARNING_ICON);
+export const infoIcon = () => cy.get(INFO_ICON);
 export const getDataElementByValue = (element) =>
   cy.get(`[data-element="${element}"]`);
 export const commonDataElementInputPreview = () =>