+++ src/components/popover-container/popover-container-test.stories.tsx
@@ -1,0 +1,37 @@
+/* eslint-disable react/prop-types */
+import React from "react";
+
+import specialCharacters from "../../__internal__/utils/argTypes/specialCharacters";
+import PopoverContainer from "./popover-container.component";
+
+export default {
+  title: "Popover Container/Test",
+  parameters: {
+    info: { disable: true },
+    chromatic: {
+      disable: true,
+    },
+  },
+  argTypes: {
+    titleSpecialCharacters: specialCharacters,
+  },
+};
+
+export const Default = ({
+  title,
+  titleSpecialCharacters,
+  open,
+}: {
+  title?: string;
+  titleSpecialCharacters?: string;
+  open: boolean;
+}) => <PopoverContainer title={title || titleSpecialCharacters} open={open} />;
+
+Default.story = {
+  name: "default",
+  args: {
+    title: "Title",
+    titleSpecialCharacters: undefined,
+    open: true,
+  },
+};